+++ esp-hal/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- RMT: `TxChannelConfig` and `RxChannelConfig` now support the builder-lite pattern (#2978)
 
 ### Fixed